requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.25.8
anyio==4.3.0
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.5
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
duckdb==0.10.2
exceptiongroup==1.2.1
faiss-cpu==1.8.0
Faker==19.13.0
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.19
langchain-anthropic==0.1.11
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.56
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.27.0
orjson==3.10.3
packaging==23.2
pandas==1.5.3
pandasai==2.0.37
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
specklepy==2.18.3
SQLAlchemy==2.0.30
streamlit==1.34.0
stringcase==1.2.0
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
